SIG: Move config to internal
