+++ go/sig/internal/config/config.go
