+++ go/sig/internal/config/config_test.go
