+++ go/sig/internal/config/testdata/01-loadfromfile.json
