+++ go/sig/internal/config/testdata/README.md
